% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfidf.R
\name{content_recommender}
\alias{content_recommender}
\title{Content Recommender}
\usage{
content_recommender(tfidf, n_threads = max(1, parallel::detectCores() - 1))
}
\arguments{
\item{tfidf}{sparse Matrix of TF-IDF}

\item{n_threads}{number of threads to use}
}
\value{
ContentRecommender
}
\description{
Recommend based on TF-IDF & cosine similarity
}
