% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid.R
\name{hybrid_recommender}
\alias{hybrid_recommender}
\title{Hybrid Recommender}
\usage{
hybrid_recommender(..., weights = NULL, n_threads = max(1,
  parallel::detectCores() - 1))
}
\arguments{
\item{n_threads}{}
}
\description{
Hybrid Recommender
}
