% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{recommend_ratings}
\alias{recommend_ratings}
\title{Recommender methods}
\usage{
recommend_ratings(recommender, test, ...)
}
\arguments{
\item{recommender}{VirtualRecommender object}

\item{test}{test matrix}

\item{...}{}
}
\description{
Recommender methods
}
