% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{popular_recommender}
\alias{popular_recommender}
\title{Popular Recommender}
\usage{
popular_recommender(train, n_threads = max(1, parallel::detectCores() - 1))
}
\arguments{
\item{n_threads}{}
}
\description{
Popular Recommender
}
