% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{random_recommender}
\alias{random_recommender}
\title{Random Recommender}
\usage{
random_recommender(train, seed = round(runif(1) * 1000), n_threads = max(1,
  parallel::detectCores() - 1))
}
\arguments{
\item{n_threads}{}
}
\description{
Random Recommender
}
