% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{test_train_split}
\alias{test_train_split}
\alias{cv_splits}
\title{Split matrix}
\usage{
test_train_split(x, train = 0.8, given = -1L)

cv_splits(x, fold = 5, given = -1)
}
\arguments{
\item{given}{}
}
\description{
Split matrix
}
