% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubcf.R
\name{ubcf_recommender}
\alias{ubcf_recommender}
\title{User Based Collaborative Filtering Recommender}
\usage{
ubcf_recommender(train, nn = 100L, n_threads = max(1,
  parallel::detectCores() - 1))
}
\arguments{
\item{n_threads}{}
}
\description{
User Based Collaborative Filtering Recommender
}
